+++ brutespray.py
@@ -1,13 +1,12 @@
-#!/usr/bin/python
+#! /usr/bin/python3
 # -*- coding: utf-8 -*-
 from argparse import RawTextHelpFormatter
 import readline, glob
 import sys, time, os
 import subprocess
-import xml.dom.minidom
+import xml.etree.ElementTree as ET
 import re
 import argparse
-import argcomplete
 import threading
 import itertools
 import tempfile
@@ -58,11 +57,13 @@ banner = colors.red + r"""
         ╚═════╝ ╚═╝  ╚═╝ ╚═════╝    ╚═╝   ╚══════╝╚══════╝╚═╝     ╚═╝  ╚═╝╚═╝  ╚═╝   ╚═╝   
                                                                                    
 """+'\n' \
-+ '\n brutespray.py v1.6.5' \
++ '\n brutespray.py v1.6.8' \
 + '\n Created by: Shane Young/@x90skysn3k && Jacob Robles/@shellfail' \
 + '\n Inspired by: Leon Johnson/@sho-luv' \
 + '\n Credit to Medusa: JoMo-Kun / Foofus Networks <jmk@foofus.net>\n' + colors.normal
 #ascii art by: Cara Pearson
+
+quiet_banner = colors.red + '~ BruteSpray ~' + colors.normal
 
 class tabCompleter(object):
 
@@ -75,47 +76,47 @@ def interactive():
     t = tabCompleter()
     singluser = ""
     if args.interactive is True:
-        print (colors.white + "\n\nWelcome to interactive mode!\n\n" + colors.normal)
-        print (colors.red + "WARNING:" + colors.white + " Leaving an option blank will leave it empty and refer to default\n\n" + colors.normal)
-        print ("Available services to brute-force:")
+        print(colors.white + "\n\nWelcome to interactive mode!\n\n" + colors.normal)
+        print(colors.red + "WARNING:" + colors.white + " Leaving an option blank will leave it empty and refer to default\n\n" + colors.normal)
+        print("Available services to brute-force:")
         for serv in services:
             srv = serv
             for prt in services[serv]:
                 iplist = services[serv][prt]
                 port = prt
                 plist = len(iplist)
-                print ("Service: " + colors.green + str(serv) + colors.normal + " on port " + colors.red + str(port) + colors.normal + " with " + colors.red + str(plist) + colors.normal + " hosts")
+                print("Service: " + colors.green + str(serv) + colors.normal + " on port " + colors.red + str(port) + colors.normal + " with " + colors.red + str(plist) + colors.normal + " hosts")
 
-        args.service = raw_input('\n' + colors.lightblue + 'Enter services you want to brute - default all (ssh,ftp,etc): ' + colors.red)
+        args.service = input('\n' + colors.lightblue + 'Enter services you want to brute - default all (ssh,ftp,etc): ' + colors.red)
 
-        args.threads = raw_input(colors.lightblue + 'Enter the number of parallel threads (default is 2): ' + colors.red)
+        args.threads = input(colors.lightblue + 'Enter the number of parallel threads (default is 2): ' + colors.red)
 
-        args.hosts = raw_input(colors.lightblue + 'Enter the number of parallel hosts to scan per service (default is 1): ' + colors.red)
+        args.hosts = input(colors.lightblue + 'Enter the number of parallel hosts to scan per service (default is 1): ' + colors.red)
 
         if args.passlist is None or args.userlist is None:
-            customword = raw_input(colors.lightblue + 'Would you like to specify a wordlist? (y/n): ' + colors.red)
+            customword = input(colors.lightblue + 'Would you like to specify a wordlist? (y/n): ' + colors.red)
         if customword == "y":
             readline.set_completer_delims('\t')
             readline.parse_and_bind("tab: complete")
             readline.set_completer(t.pathCompleter)
             if args.userlist is None and args.username is None:
-                args.userlist = raw_input(colors.lightblue + 'Enter a userlist you would like to use: ' + colors.red)
+                args.userlist = input(colors.lightblue + 'Enter a userlist you would like to use: ' + colors.red)
                 if args.userlist == "":
                     args.userlist = None
             if args.passlist is None and args.password is None:
-                args.passlist = raw_input(colors.lightblue + 'Enter a passlist you would like to use: ' + colors.red)
+                args.passlist = input(colors.lightblue + 'Enter a passlist you would like to use: ' + colors.red)
                 if args.passlist == "":
                     args.passlist = None
 
         if args.username is None or args.password is None: 
-            singluser = raw_input(colors.lightblue + 'Would to specify a single username or password (y/n): ' + colors.red)
+            singluser = input(colors.lightblue + 'Would to specify a single username or password (y/n): ' + colors.red)
         if singluser == "y":
             if args.username is None and args.userlist is None:
-                args.username = raw_input(colors.lightblue + 'Enter a username: ' + colors.red)
+                args.username = input(colors.lightblue + 'Enter a username: ' + colors.red)
                 if args.username == "":
                     args.username = None
             if args.password is None and args.passlist is None:
-                args.password = raw_input(colors.lightblue + 'Enter a password: ' + colors.red)
+                args.password = input(colors.lightblue + 'Enter a password: ' + colors.red)
                 if args.password == "":
                     args.password = None
 
@@ -126,7 +127,7 @@ def interactive():
         if args.hosts == "":
             args.hosts = "1"
 
-    print colors.normal
+    print(colors.normal)
 
 NAME_MAP = {"ms-sql-s": "mssql",
             "microsoft-ds": "smbnt",
@@ -221,46 +222,20 @@ def make_dic_xml(filename):
                  'vnc','imap','imaps','nntp','pcanywheredata','pop3','pop3s',
                  'exec','login','microsoft-ds','smtp','smtps','submission',
                  'svn','iss-realsecure','snmptrap','snmp']
-    doc = xml.dom.minidom.parse(filename)
-    for host in doc.getElementsByTagName("host"):
-        try:
-            address = host.getElementsByTagName("address")[0]
-            ip = address.getAttribute("addr")
-            eip = ip.encode("utf8")
-            iplist = eip.split(',')
-        except:
-            # move to the next host
-            continue
-        try:
-            status = host.getElementsByTagName("status")[0]
-            state = status.getAttribute("state")
-        except:
-            state = ""
-        try:
-            ports = host.getElementsByTagName("ports")[0]
-            ports = ports.getElementsByTagName("port")
-        except:
-            continue
-
-        for port in ports:
-            pn = port.getAttribute("portid")
-            state_el = port.getElementsByTagName("state")[0]
-            state = state_el.getAttribute("state")
-            if state == "open":
+    tree = ET.parse(args.file)
+    root = tree.getroot()
+    for host in root.iter('host'):
+        ipaddr = host.find('address').attrib['addr']
+        for port in host.iter('port'):
+            cstate = port.find('state').attrib['state']
+            if cstate == "open":
                 try:
-                    service = port.getElementsByTagName("service")[0]
-                    port_name = service.getAttribute("name")
+                    name = port.find('service').attrib['name']
+                    tmp_port = port.attrib['portid']
+                    iplist = ipaddr.split(',')
                 except:
-                    service = ""
-                    port_name = ""
-                    product_descr = ""
-                    product_ver = ""
-                    product_extra = ""
-                name = port_name.encode("utf-8")
-                tmp_port = pn.encode("utf-8")
-
+                    continue
                 if name in supported:
-
                     name = NAME_MAP.get(name, name)
                     if name in services:
                         if tmp_port in services[name]:
@@ -269,12 +244,43 @@ def make_dic_xml(filename):
                             services[name][tmp_port] = iplist
                     else:
                         services[name] = {tmp_port:iplist}
+
     loading = True
 
+def make_dic_json():
+    global loading
+    global services
+
+    supported = ['ssh','ftp','postgres','telnet','mysql','ms-sql-s','shell',
+                 'vnc','imap','imaps','nntp','pcanywheredata','pop3','pop3s',
+                 'exec','login','microsoft-ds','smtp', 'smtps','submission',
+                 'svn','iss-realsecure','snmptrap','snmp']
+
+    with open(args.file, "r") as jsonlines_file:
+        for line in jsonlines_file:
+            data = json.loads(line)
+            try:
+                host, port, name = data["host"], data["port"], data["service"]
+                
+                if name in supported:
+                    name = NAME_MAP.get(name, name) 
+                    if name not in services:
+                        services[name] = {}
+                    if port not in services[name]:
+                        services[name][port] = []
+                    if host not in services[name][port]:
+                        services[name][port].append(host)
+            except KeyError as e:
+                sys.stderr.write("\n[!] Field: " + str(e) + "is missing")
+                sys.stderr.write("\n[!] Please provide the json fields. ")               
+                continue
+    loading = True 
 
 def brute(service,port,fname,output):
-    if args.userlist is None and args.username is None and args.combo is None:
-        userlist = 'wordlist/'+service+'/user'
+    if args.userlist is None and args.username is None:
+        userlist = '/usr/share/brutespray/wordlist/'+service+'/user'
+        if not os.path.exists(userlist):
+            userlist = 'wordlist/'+service+'/user'
         uarg = '-U'
     elif args.userlist:
         userlist = args.userlist
@@ -287,8 +293,10 @@ def brute(service,port,fname,output):
         uarg = "-C"
 
 
-    if args.passlist is None and args.password is None and args.combo is None:
-        passlist = 'wordlist/'+service+'/password'
+    if args.passlist is None and args.password is None:
+        passlist = '/usr/share/brutespray/wordlist/'+service+'/password'
+        if not os.path.exists(passlist):
+            passlist = 'wordlist/'+service+'/password'
         parg = '-P'
     elif args.passlist:
         passlist = args.passlist
@@ -318,12 +326,12 @@ def brute(service,port,fname,output):
     
  
     for line in iter(p.stdout.readline, b''):
-        print line,
+        print(line.decode('utf-8').strip('\n'))
         sys.stdout.flush()
         time.sleep(0.0001)
-        if 'SUCCESS' in line:
+        if 'SUCCESS' in line.decode('utf-8'):
             f = open(output_file, 'a')
-            f.write(out + line)
+            f.write(out + line.decode('utf-8'))
             f.close()
 
 def animate():
@@ -346,6 +354,28 @@ def loading():
         sys.stdout.flush()
         time.sleep(0.01)
 
+def getInput(filename):
+    in_format = None
+    with open(filename) as f:
+        line = f.readlines()
+        if filename.endswith("gnmap"):
+            in_format = "gnmap"
+        if filename.endswith("json"):
+            in_format = "json"
+        if filename.endswith("xml"):
+            in_format = "xml"
+        if '{' in line[0]:
+            in_format = "json"
+        if '# Nmap' in line[0] and not 'Nmap' in line[1]:
+            in_format = "gnmap"
+        if '<?xml ' in line[0]:
+            in_format = "xml"
+        if in_format is None:
+            print('File is not correct format!\n')
+            sys.exit(0)
+
+    return in_format 
+
 def parse_args():
 
     parser = argparse.ArgumentParser(formatter_class=RawTextHelpFormatter, description=\
@@ -354,6 +384,7 @@ def parse_args():
 
     menu_group = parser.add_argument_group(colors.lightblue + 'Menu Options' + colors.normal)
 
+    #menu_group.add_argument('-f', '--file', help="GNMAP or XML file to parse", required=False, default=None)
     menu_group.add_argument('-f', '--file', help="GNMAP, JSON or XML file to parse", required=False, default=None)
     menu_group.add_argument('-o', '--output', help="Directory containing successful attempts", default="brutespray-output")
     menu_group.add_argument('-s', '--service', help="specify service to attack", default="all")
@@ -367,8 +398,8 @@ def parse_args():
     menu_group.add_argument('-c', '--continuous', help="keep brute-forcing after success", default=False, action='store_true')
     menu_group.add_argument('-i', '--interactive', help="interactive mode", default=False, action='store_true')    
     menu_group.add_argument('-m', '--modules', help="dump a list of available modules to brute", default=False, action='store_true')    
+    menu_group.add_argument('-q', '--quiet', help="supress banner", default=False, action='store_true')   
 
-    argcomplete.autocomplete(parser)
     args = parser.parse_args()
 
     if args.file is None and args.modules is False:
@@ -402,8 +433,12 @@ def detectFormat(filename):
     return in_format 
 
 if __name__ == "__main__":
-    print(banner)
     args = parse_args()
+    
+    if args.quiet == False:
+        print(banner)
+    else:
+        print(quiet_banner)
 
     supported = ['ssh','ftp','telnet','vnc','mssql','mysql','postgresql','rsh',
                 'imap','nntp','pcanywhere','pop3',
@@ -412,9 +447,9 @@ if __name__ == "__main__":
     #temporary directory for ip addresses
 
     if args.modules is True:
-        print (colors.lightblue + "Supported Services:\n" + colors.green)
-        print ('\n'.join(supported))
-        print (colors.normal + "\n" )
+        print(colors.lightblue + "Supported Services:\n" + colors.green)
+        print(('\n'.join(supported)))
+        print(colors.normal + "\n") 
     try:
         tmppath = tempfile.mkdtemp(prefix="brutespray-tmp")
     except:
@@ -448,14 +483,16 @@ if __name__ == "__main__":
             
             t = threading.Thread(target=loading)
             t.start()
+            in_format = getInput(args.file)
             {
                 "gnmap": make_dic_gnmap,
                 "xml":   make_dic_xml,
                 "json":  make_dic_json
-            }[in_format](args.file) 
+            }[in_format]()
         except:
-            print("\nFormat failed ! ")
-            exit(3)
+            print("\nFormat failed!\n")
+            loading = True
+            sys.exit(0)
 
         if args.interactive is True:
             interactive()
@@ -463,9 +500,9 @@ if __name__ == "__main__":
         animate()
 
         if services == {}:
-            print ("\nNo brutable services found.\n Please check your Nmap file.")
+            print("\nNo brutable services found.\n Please check your Nmap file.")
     else:
-        print ("\nError loading file, please check your filename.")
+        print("\nError loading file, please check your filename.")
 
     to_scan = args.service.split(',')